project/requirements.txt
streamlit>=1.24.0
networkx>=3.1
matplotlib>=3.7.1
numpy>=1.24.3
openai>=1.3.0
python-dotenv>=1.0.0